Update arduinoHSV.c
Switch to unsigned int arithmetic to avoid overflows.

+++ arduinoHSV.c
@@ -5,11 +5,11 @@ byte BlueLight;
 void setLedColorHSV(byte h, byte s, byte v) {
   // this is the algorithm to convert from RGB to HSV
   h = (h * 192) / 256;  // 0..191
-  byte i = h / 32;   // We want a value of 0 thru 5
-  byte f = (h % 32) * 8;   // 'fractional' part of 'i' 0..248 in jumps
+  unsigned int i = h / 32;   // We want a value of 0 thru 5
+  unsigned int f = (h % 32) * 8;   // 'fractional' part of 'i' 0..248 in jumps
 
-  byte sInv = 255 - s;  // 0 -> 0xff, 0xff -> 0
-  byte fInv = 255 - f;  // 0 -> 0xff, 0xff -> 0
+  unsigned int sInv = 255 - s;  // 0 -> 0xff, 0xff -> 0
+  unsigned int fInv = 255 - f;  // 0 -> 0xff, 0xff -> 0
   byte pv = v * sInv / 256;  // pv will be in range 0 - 255
   byte qv = v * (256 - s * f / 256) / 256;
   byte tv = v * (256 - s * fInv / 256) / 256;
